Refactor server setup functions to a new `setup` module
